+++ src/protocols/sql_custom.h
@@ -34,7 +34,7 @@ class SQL_CUSTOM: public AbstractProtocol
 {
 	public:
 		bool init(AbstractExt *extension, const std::string &database_id, const std::string init_str);
-		bool callProtocol(std::string input_str, std::string &result, const bool async_method, const unsigned long unique_id=1);
+		bool callProtocol(std::string input_str, std::string &result, const bool async_method, const unsigned int unique_id=1);
 		
 	private:
 		AbstractExt::DBConnectionInfo *database_ptr;
