+++ src/protocols/steam_v2.cpp
@@ -54,7 +54,7 @@ bool STEAM_V2::isNumber(const std::string &input_str)
 }
 
 
-bool STEAM_V2::callProtocol(std::string input_str, std::string &result, const bool async_method, const unsigned long unique_id)
+bool STEAM_V2::callProtocol(std::string input_str, std::string &result, const bool async_method, const unsigned int unique_id)
 {
 	#ifdef DEBUG_TESTING
 		extension_ptr->console->info("extDB2: STEAM_V2: Trace: Input: {0}", input_str);
